backend/accounts/services/jwt_token.py
from rest_framework_simplejwt.tokens import RefreshToken
from django.contrib.auth.models import AbstractBaseUser

def GENERATE_JWT_TOKEN(user: AbstractBaseUser) -> dict[str, str]:
    refresh = RefreshToken.for_user(user)
    return {
        'refresh_token': str(refresh),
        'access_token': str(refresh.access_token),
    }
