backend/requirements/dev.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
phonenumbers==9.0.7
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
attrs==25.3.0
django-schema-viewer==0.5.3
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
uritemplate==4.2.0
